backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argcomplete==3.6.2
asgiref==3.8.1
babel==2.17.0
beautifulsoup4==4.13.3
boto3==1.37.29
botocore==1.37.29
cachetools==4.2.4
certifi==2025.1.31
cfn-flip==1.3.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
courlan==1.3.2
dateparser==1.2.1
distro==1.9.0
Django==5.1.7
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.15.2
durationpy==0.9
google-api-core==1.34.1
google-auth==1.35.0
google-cloud-core==1.7.3
google-cloud-translate==2.0.1
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.49.0rc1
gunicorn==23.0.0
h11==0.14.0
hjson==3.1.0
htmldate==1.9.3
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
jmespath==1.0.1
jusText==3.0.2
kappa==0.6.0
lxml==5.3.2
lxml_html_clean==0.4.1
MarkupSafe==3.0.2
openai==1.68.2
packaging==24.2
placebo==0.9.0
protobuf==3.20.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
s3transfer==0.11.4
setuptools==78.0.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
text-unidecode==1.3
tld==0.13
toml==0.10.2
tqdm==4.67.1
trafilatura==2.0.0
troposphere==4.9.0
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
zappa==0.59.0
